Move terminalPanel, electron-browser->browser
Part of #69115
