+++ src/tsconfig.strictNullChecks.json
@@ -281,6 +281,7 @@
 		"./vs/workbench/contrib/terminal/browser/terminalActions.ts",
 		"./vs/workbench/contrib/terminal/browser/terminalFindWidget.ts",
 		"./vs/workbench/contrib/terminal/browser/terminalLinkHandler.ts",
+		"./vs/workbench/contrib/terminal/browser/terminalPanel.ts",
 		"./vs/workbench/contrib/terminal/browser/terminalQuickOpen.ts",
 		"./vs/workbench/contrib/terminal/browser/terminalTab.ts",
 		"./vs/workbench/contrib/terminal/browser/terminalWidgetManager.ts",
@@ -292,7 +293,6 @@
 		"./vs/workbench/contrib/terminal/common/terminalService.ts",
 		"./vs/workbench/contrib/terminal/electron-browser/terminalConfigHelper.ts",
 		"./vs/workbench/contrib/terminal/electron-browser/terminalInstance.ts",
-		"./vs/workbench/contrib/terminal/electron-browser/terminalPanel.ts",
 		"./vs/workbench/contrib/terminal/electron-browser/terminalProcessManager.ts",
 		"./vs/workbench/contrib/terminal/node/terminal.ts",
 		"./vs/workbench/contrib/terminal/node/terminalCommandTracker.ts",
